+++ CHANGELOG.md
@@ -69,6 +69,8 @@ and this project adheres to
 
 ### Changed
 
+- Prevent to enroll to not listed course runs related to an order awaiting
+  signature of a contract by the learner
 - Use Invoice.recipient_address to populate Contract address context
 - Link Invoice to Address object
 - Update psycopg to version 3